remove redundant header + some fix by clangd recommend
