+++ src/include/miopen/multimarginloss/problem_description.hpp
@@ -30,8 +30,6 @@
 #include <miopen/activ.hpp>
 #include <miopen/problem_description_base.hpp>
 #include <miopen/tensor.hpp>
-#include <cassert>
-#include <string>
 
 namespace miopen {
