+++ src/kernels/MIOpenMultiMarginLoss.cpp
