+++ test/gtest/multimarginloss.hpp
@@ -24,16 +24,13 @@
  *
  *******************************************************************************/
 
-#include "../driver/tensor_driver.hpp"
 #include "cpu_multimarginloss.hpp"
 #include "get_handle.hpp"
-#include "random.hpp"
 #include "tensor_holder.hpp"
 #include "verify.hpp"
 #include <gtest/gtest.h>
 #include <miopen/miopen.h>
 #include <miopen/multimarginloss.hpp>
-#include <numeric>
 
 struct MultiMarginLossTestCase
 {
@@ -42,26 +39,6 @@ struct MultiMarginLossTestCase
     miopenLossReductionMode_t reduction_mode;
     long p;
 };
-
-std::vector<MultiMarginLossTestCase> MultiMarginLossTestConfigs()
-{
-    // clang-format off
-    return {
-    {{22, 12}, true, MIOPEN_LOSS_REDUCTION_MEAN, 1}, 
-    {{22, 12}, false, MIOPEN_LOSS_REDUCTION_SUM, 1}, 
-    {{22, 12}, true, MIOPEN_LOSS_REDUCTION_NONE, 1}, 
-    {{9456, 13}, false, MIOPEN_LOSS_REDUCTION_MEAN, 2 }, 
-    {{9456, 13}, true, MIOPEN_LOSS_REDUCTION_SUM, 2 }, 
-    {{9456, 13}, false, MIOPEN_LOSS_REDUCTION_NONE, 2 }, 
-    {{543210, 7}, true, MIOPEN_LOSS_REDUCTION_MEAN, 2 }, 
-    {{543210, 7}, false, MIOPEN_LOSS_REDUCTION_SUM, 2 }, 
-    {{543210, 7}, true, MIOPEN_LOSS_REDUCTION_NONE, 2 }, 
-    {{3995776, 6}, true, MIOPEN_LOSS_REDUCTION_MEAN, 1 }, 
-    {{3995776, 6}, true, MIOPEN_LOSS_REDUCTION_SUM, 1 }, 
-    {{3995776, 6}, true, MIOPEN_LOSS_REDUCTION_NONE, 1 }, 
-    };
-    // clang-format on
-}
 
 template <typename T = float>
 struct MultiMarginLossForwardTest : public ::testing::TestWithParam<MultiMarginLossTestCase>
